+++ src/parser/api/cache.ts
@@ -1,7 +1,9 @@
+import { cache as code } from '../block/codeblock';
 import { cache as math } from '../inline/math';
 import { cache as media } from '../inline/media';
 
 export const caches = {
+  code,
   math,
   media,
 };