+++ src/parser/inline/media.ts
@@ -11,7 +11,7 @@ import { Cache } from 'spica/cache';
 
 const url = html('a');
 
-export const cache = new Cache<string, HTMLElement>(10);
+export const cache = new Cache<string, HTMLElement>(10); // for rerendering in editing
 
 export const media: MediaParser = lazy(() => creator(10, bind(fmap(open(
   '!',