+++ src/renderer/render/code.ts
@@ -1,6 +1,9 @@
 import * as Prism from 'prismjs';
+import { Cache } from 'spica/cache';
 
-export function code(target: HTMLElement): void {
+export function code(target: HTMLElement, cache?: Cache<string, HTMLElement>): void {
   assert(target.children.length === 0);
-  requestAnimationFrame(() => Prism.highlightElement(target, false));
+  const source = target.textContent!;
+  Prism.highlightElement(target, false, () =>
+    void cache?.set(`${target.getAttribute('data-lang') || ''}\n${source}`, target.cloneNode(true)));
 }